+++ include/app/WaterLogFileStore.h
@@ -28,7 +28,7 @@ public:
 
     bool begin();
     bool append(const WaterLogRecord& record);
-    std::size_t rewriteBootRelativeTimes(std::uint16_t bootId, std::uint32_t bootStartRealSec);
+    std::size_t rewriteBootRelativeTimes(std::uint32_t bootId, std::uint32_t bootStartRealSec);
     std::size_t readPage(std::size_t pageIndex,
                          std::uint16_t pageSize,
                          WaterLogRecord* output,
