+++ include/web/FaucetWeb.h
@@ -11,7 +11,7 @@ class WaterLogReader;
 struct SystemConfig;
 
 using FaucetNowSeconds = std::uint32_t (*)();
-using FaucetBootId = std::uint16_t (*)();
+using FaucetBootId = std::uint32_t (*)();
 using FaucetApplySettings = void (*)(const SystemConfig&);
 
 struct FaucetWebContext {
